Fix for D-Pad Right overlapping with C-buttons (Extrems, vaguerant, emukidid)
Reference: https://github.com/emukidid/Wii64/commit/192b7c022c3d85f4425ee35b837a562558d1308a

+++ gc_input/controller-Classic.c
@@ -99,7 +99,7 @@ static button_t menu_combos[] = {
 
 static unsigned int getButtons(classic_ctrl_t* controller)
 {
-	unsigned int b = (unsigned)controller->btns;
+	unsigned int b = (unsigned short)controller->btns;
 	s8 stickX      = getStickValue(&controller->ljs, STICK_X, 7);
 	s8 stickY      = getStickValue(&controller->ljs, STICK_Y, 7);
 	s8 substickX   = getStickValue(&controller->rjs, STICK_X, 7);
